content_planner_agent/config/settings.py
"""
Configuration settings for the Content Planner Agent.
"""
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# LLM Configuration
MODEL_NAME = "claude-3-sonnet-20240229"
TEMPERATURE = 0.7
ANTHROPIC_API_KEY = os.getenv("ANTHROPIC_API_KEY")

# Database Configuration
DB_CONNECTION_STRING = os.getenv("DB_CONNECTION_STRING", "mongodb://localhost:27017")
DB_NAME = os.getenv("DB_NAME", "content_planner")
GUIDELINES_COLLECTION = "guidelines"
SESSIONS_COLLECTION = "sessions"

# System prompt for the agent
SYSTEM_PROMPT = """
You are a helpful assistant specialized in creating content guidelines for automated social media posting.

Your goal is to help the user create a comprehensive content guideline by:
1. Understanding their brand voice, audience, and goals
2. Suggesting content themes and topics
3. Recommending posting frequency and best practices
4. Providing examples of effective posts
5. Iteratively refining the guideline based on user feedback

As you work with the user, maintain a draft of the content guideline that gets more detailed over time.
When the user is satisfied with the guideline, they can save the final version.

When you have new information to add to the guideline, include a section in your response like this:

GUIDELINE UPDATE:
[Your updated guideline content here]
END GUIDELINE UPDATE

Format your responses using Markdown:
- Use **bold** for emphasis
- Use bullet points and numbered lists for structured information
- Use headings (## and ###) for sections
- Use code blocks for examples when appropriate

Be concise, helpful, and focus on creating practical, actionable guidelines.
""" 