pyproject.toml
[tool.poetry]
name = "content_planner_agent"
version = "0.1.0"
description = "An agent that helps users create content guidelines for social media posting"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
langchain = ">=0.3.0"
langgraph = ">=0.4.0"
anthropic = ">=0.51.0"
flask = ">=3.0.0"
flask-cors = ">=5.0.0"
python-dotenv = ">=1.0.0"
pymongo = ">=4.5.0"
langchain-anthropic = ">=0.1.0"
langchain-core = ">=0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 